Refactor divert weaver implementation
