+++ cmd/deploy/deploy.go
@@ -32,9 +32,9 @@ import (
 	"github.com/okteto/okteto/pkg/cmd/pipeline"
 	"github.com/okteto/okteto/pkg/cmd/stack"
 	"github.com/okteto/okteto/pkg/constants"
+	"github.com/okteto/okteto/pkg/divert/weaver"
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/format"
-	"github.com/okteto/okteto/pkg/k8s/diverts"
 	"github.com/okteto/okteto/pkg/k8s/ingresses"
 	"github.com/okteto/okteto/pkg/k8s/kubeconfig"
 	oktetoLog "github.com/okteto/okteto/pkg/log"
@@ -532,24 +532,8 @@ func (dc *DeployCommand) deployDivert(ctx context.Context, opts *Options) error 
 		return err
 	}
 
-	cache, err := diverts.InitDivertCache(ctx, opts.Manifest, c)
-	if err != nil {
-		return err
-	}
-
-	for name, in := range cache.DivertIngresses {
-		select {
-		case <-ctx.Done():
-			oktetoLog.Infof("deployDivert context cancelled")
-			return ctx.Err()
-		default:
-			oktetoLog.Spinner(fmt.Sprintf("Diverting ingress %s/%s...", in.Namespace, in.Name))
-			if err := diverts.DivertIngress(ctx, opts.Manifest, name, cache, c); err != nil {
-				return err
-			}
-		}
-	}
-	return diverts.CreateDivertCRD(ctx, opts.Manifest)
+	driver := weaver.New(c, opts.Manifest)
+	return driver.Deploy(ctx)
 }
 
 func (dc *DeployCommand) deployEndpoints(ctx context.Context, opts *Options) error {
